# tests/fixtures/platform/openapi-v1.json
{"components": {"schemas": {"AgentCreateRequest": {"additionalProperties": false, "properties": {"revision_id": {"title": "Revision Id", "type": "string"}}, "required": ["revision_id"], "title": "AgentCreateRequest", "type": "object"}, "BudgetSpec": {"additionalProperties": false, "properties": {"max_cases": {"anyOf": [{"minimum": 1.0, "type": "integer"}, {"type": "null"}], "title": "Max Cases"}, "max_cost": {"anyOf": [{"minimum": 0.0, "type": "number"}, {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "type": "string"}, {"type": "null"}], "title": "Max Cost"}, "max_tokens": {"anyOf": [{"minimum": 0.0, "type": "integer"}, {"type": "null"}], "title": "Max Tokens"}}, "title": "BudgetSpec", "type": "object"}, "ErrorEnvelope": {"additionalProperties": false, "properties": {"error": {"additionalProperties": false, "properties": {"code": {"type": "string"}, "details": {"type": "object"}, "message": {"type": "string"}, "request_id": {"format": "uuid", "type": "string"}}, "required": ["code", "message", "details", "request_id"], "type": "object"}}, "required": ["error"], "type": "object"}, "ModelCreateRequest": {"additionalProperties": false, "properties": {"endpoint": {"maxLength": 2048, "minLength": 1, "title": "Endpoint", "type": "string"}, "model": {"maxLength": 256, "minLength": 1, "title": "Model", "type": "string"}, "name": {"maxLength": 128, "minLength": 1, "title": "Name", "type": "string"}, "parameters": {"$ref": "#/components/schemas/ModelParameters"}, "pricing": {"$ref": "#/components/schemas/ModelPricing"}, "provider": {"maxLength": 64, "minLength": 1, "title": "Provider", "type": "string"}}, "required": ["name", "provider", "endpoint", "model"], "title": "ModelCreateRequest", "type": "object"}, "ModelParameters": {"additionalProperties": false, "properties": {"frequency_penalty": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Frequency Penalty"}, "max_output_tokens": {"anyOf": [{"minimum": 0.0, "type": "integer"}, {"type": "null"}], "title": "Max Output Tokens"}, "max_tokens": {"anyOf": [{"minimum": 0.0, "type": "integer"}, {"type": "null"}], "title": "Max Tokens"}, "presence_penalty": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Presence Penalty"}, "seed": {"anyOf": [{"minimum": 0.0, "type": "integer"}, {"type": "null"}], "title": "Seed"}, "stop": {"anyOf": [{"type": "string"}, {"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Stop"}, "temperature": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Temperature"}, "top_p": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Top P"}}, "title": "ModelParameters", "type": "object"}, "ModelPricing": {"additionalProperties": false, "properties": {"input_per_million": {"anyOf": [{"minimum": 0.0, "type": "number"}, {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "type": "string"}, {"type": "null"}], "title": "Input Per Million"}, "output_per_million": {"anyOf": [{"minimum": 0.0, "type": "number"}, {"pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$", "type": "string"}, {"type": "null"}], "title": "Output Per Million"}}, "title": "ModelPricing", "type": "object"}, "ReviewCreateRequest": {"additionalProperties": false, "properties": {"label": {"maxLength": 128, "minLength": 1, "title": "Label", "type": "string"}, "notes": {"default": "", "maxLength": 4096, "title": "Notes", "type": "string"}}, "required": ["label"], "title": "ReviewCreateRequest", "type": "object"}, "TaskCreateRequest": {"additionalProperties": false, "properties": {"agent_revision_id": {"title": "Agent Revision Id", "type": "string"}, "benchmark_id": {"title": "Benchmark Id", "type": "string"}, "budget": {"$ref": "#/components/schemas/BudgetSpec"}, "case_ids": {"items": {"type": "string"}, "minItems": 1, "title": "Case Ids", "type": "array"}, "name": {"maxLength": 256, "minLength": 1, "title": "Name", "type": "string"}}, "required": ["name", "benchmark_id", "agent_revision_id", "case_ids"], "title": "TaskCreateRequest", "type": "object"}, "TaskCreateResponse": {"additionalProperties": false, "properties": {"id": {"title": "Id", "type": "string"}, "run_id": {"title": "Run Id", "type": "string"}, "status": {"title": "Status", "type": "string"}, "task_spec_sha256": {"title": "Task Spec Sha256", "type": "string"}}, "required": ["id", "run_id", "status", "task_spec_sha256"], "title": "TaskCreateResponse", "type": "object"}}}, "info": {"title": "SecRL Lite API", "version": "1.0.0"}, "openapi": "3.1.0", "paths": {"/api/v1/agents": {"get": {"operationId": "list_agents_api_v1_agents_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"additionalProperties": true, "type": "object"}, "title": "Response List Agents Api V1 Agents Get", "type": "array"}}}, "description": "Successful Response"}}, "summary": "List Agents", "tags": ["agents"]}, "post": {"operationId": "create_agent_api_v1_agents_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AgentCreateRequest"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Create Agent Api V1 Agents Post", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Create Agent", "tags": ["agents"]}}, "/api/v1/agents/{id}:check": {"post": {"operationId": "check_agent_api_v1_agents__id__check_post", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Check Agent Api V1 Agents  Id  Check Post", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Check Agent", "tags": ["agents"]}}, "/api/v1/artifacts/{id}": {"get": {"operationId": "download_artifact_api_v1_artifacts__id__get", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Download Artifact", "tags": ["artifacts"]}}, "/api/v1/artifacts/{id}/metadata": {"get": {"operationId": "artifact_metadata_api_v1_artifacts__id__metadata_get", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Artifact Metadata Api V1 Artifacts  Id  Metadata Get", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Artifact Metadata", "tags": ["artifacts"]}}, "/api/v1/attributions/{id}/reviews": {"post": {"operationId": "create_review_api_v1_attributions__id__reviews_post", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewCreateRequest"}}}, "required": true}, "responses": {"200": {"content": {"application/json": {"schema": {}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Create Review", "tags": ["analysis"]}}, "/api/v1/auth/login": {"post": {"operationId": "login_api_v1_auth_login_post", "responses": {"200": {"content": {"application/json": {"schema": {}}}, "description": "Successful Response"}}, "summary": "Login", "tags": ["auth"]}}, "/api/v1/auth/logout": {"post": {"operationId": "logout_api_v1_auth_logout_post", "responses": {"204": {"description": "Successful Response"}}, "summary": "Logout", "tags": ["auth"]}}, "/api/v1/benchmarks": {"get": {"operationId": "list_benchmarks_api_v1_benchmarks_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"additionalProperties": true, "type": "object"}, "title": "Response List Benchmarks Api V1 Benchmarks Get", "type": "array"}}}, "description": "Successful Response"}}, "summary": "List Benchmarks", "tags": ["benchmarks"]}}, "/api/v1/compare": {"get": {"operationId": "compare_api_v1_compare_get", "parameters": [{"in": "query", "name": "left", "required": true, "schema": {"title": "Left", "type": "string"}}, {"in": "query", "name": "right", "required": true, "schema": {"title": "Right", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Compare Api V1 Compare Get", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Compare", "tags": ["compare"]}}, "/api/v1/health": {"get": {"operationId": "health_api_v1_health_get", "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "title": "Response Health Api V1 Health Get", "type": "object"}}}, "description": "Successful Response"}}, "summary": "Health", "tags": ["health"]}}, "/api/v1/models": {"get": {"operationId": "list_models_api_v1_models_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"additionalProperties": true, "type": "object"}, "title": "Response List Models Api V1 Models Get", "type": "array"}}}, "description": "Successful Response"}}, "summary": "List Models", "tags": ["models"]}, "post": {"operationId": "create_model_api_v1_models_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModelCreateRequest"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Create Model Api V1 Models Post", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Create Model", "tags": ["models"]}}, "/api/v1/runs/{id}": {"get": {"operationId": "get_run_api_v1_runs__id__get", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Get Run Api V1 Runs  Id  Get", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Get Run", "tags": ["runs"]}}, "/api/v1/runs/{id}/analysis": {"get": {"operationId": "get_analysis_api_v1_runs__id__analysis_get", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Get Analysis", "tags": ["analysis"]}}, "/api/v1/runs/{id}/cases": {"get": {"operationId": "list_run_cases_api_v1_runs__id__cases_get", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"items": {"additionalProperties": true, "type": "object"}, "title": "Response List Run Cases Api V1 Runs  Id  Cases Get", "type": "array"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "List Run Cases", "tags": ["runs"]}}, "/api/v1/runs/{id}/cases/{case_id}:retry": {"post": {"operationId": "retry_case_api_v1_runs__id__cases__case_id__retry_post", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}, {"in": "path", "name": "case_id", "required": true, "schema": {"title": "Case Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Retry Case Api V1 Runs  Id  Cases  Case Id  Retry Post", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Retry Case", "tags": ["runs"]}}, "/api/v1/runs/{id}:analyze": {"post": {"operationId": "analyze_run_api_v1_runs__id__analyze_post", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Analyze Run", "tags": ["analysis"]}}, "/api/v1/runs/{id}:cancel": {"post": {"operationId": "cancel_run_api_v1_runs__id__cancel_post", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Cancel Run Api V1 Runs  Id  Cancel Post", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Cancel Run", "tags": ["runs"]}}, "/api/v1/runs/{id}:pause": {"post": {"operationId": "pause_run_api_v1_runs__id__pause_post", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Pause Run Api V1 Runs  Id  Pause Post", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Pause Run", "tags": ["runs"]}}, "/api/v1/runs/{id}:resume": {"post": {"operationId": "resume_run_api_v1_runs__id__resume_post", "parameters": [{"in": "path", "name": "id", "required": true, "schema": {"title": "Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": true, "title": "Response Resume Run Api V1 Runs  Id  Resume Post", "type": "object"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Resume Run", "tags": ["runs"]}}, "/api/v1/tasks": {"get": {"operationId": "list_tasks_api_v1_tasks_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"additionalProperties": true, "type": "object"}, "title": "Response List Tasks Api V1 Tasks Get", "type": "array"}}}, "description": "Successful Response"}}, "summary": "List Tasks", "tags": ["tasks"]}, "post": {"operationId": "create_task_api_v1_tasks_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskCreateRequest"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskCreateResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Validation Error"}}, "summary": "Create Task", "tags": ["tasks"]}}}}